Added to EvtDeviceAdd
